Slept on it, found time before deadline. Refactored tests and moved the datasource to a new class to make it interchangable.
