+++ jobs/Backend/Task/IExchangeRateDataSource.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ExchangeRateUpdater
+{
+    public interface IExchangeRateDataSource
+    {
+        Task<string> GetExchangeRateDataAsync(DateTime date);
+    }
+}
